going to put more work into this later today in order to get something dr roche can work on over break

+++ headers/stdlib.h
@@ -1,3 +1,3 @@
 #define _stdlib_inclusion
-typedef int size_t;
+typedef unsigned long size_t;
 void * NULL = 0;
